Completed
The Page name is FOODINFINI. In this page we use HTML5 and CSS3 to create this website and this page is responsive to by the using Media Quarries. In this Website contain Navbar :- Home, Services, About Us and Contact Us. Then the content all about foods , Our Clients and Contact Us. The food section contain the Back Ground image with description and button on the image and the boxes with food image description about food. Client Section contain back ground image with clients logo on the image. Next contact section there is the input field with name , email, Phone Number and message(text area). And lastly the footer section with Copyright.